pyproject.toml
[project]
name = "mvid"
version = "0.1.0"
description = "Simple video reading and writing"
readme = "README.md"
authors = [
    { name = "Adam Alcolado", email = "adam.alcolado@mtl.ai" }
]
requires-python = ">=3.13"
dependencies = [
    "av",
    "numpy"
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "tqdm",
    "pillow"
]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
